Add _BackgroundShapeModifier and CSS impl for Rectangle bounds
