+++ Sources/TokamakStaticHTML/Modifiers/ViewModifier.swift
@@ -44,45 +44,3 @@ extension _ZIndexModifier: DOMViewModifier {
     ["style": "z-index: \(index);"]
   }
 }
-
-extension _BackgroundStyleModifier: DOMViewModifier {
-  public var isOrderDependent: Bool { true }
-  public var attributes: [HTMLAttribute: String] {
-    if let resolved = style.resolve(
-      for: .resolveStyle(levels: 0..<1),
-      in: environment,
-      role: .fill
-    ) {
-      if case let .foregroundMaterial(rgba, material) = resolved {
-        let blur: (opacity: Double, radius: Double)
-        switch material {
-        case .ultraThin:
-          blur = (0.2, 20)
-        case .thin:
-          blur = (0.4, 25)
-        case .regular:
-          blur = (0.5, 30)
-        case .thick:
-          blur = (0.6, 40)
-        case .ultraThick:
-          blur = (0.6, 50)
-        }
-        return [
-          "style":
-            """
-            background-color: rgba(\(rgba.red * 255), \(rgba.green * 255), \(rgba
-              .blue * 255), \(blur
-              .opacity));
-            -webkit-backdrop-filter: blur(\(blur.radius)px);
-            backdrop-filter: blur(\(blur.radius)px);
-            """,
-        ]
-      } else if let color = resolved.color(at: 0) {
-        return [
-          "style": "background-color: \(color.cssValue(environment))",
-        ]
-      }
-    }
-    return [:]
-  }
-}
